qid: Reorganized the service files into entities
